LPS-52679 Rename so everything is consistent
