+++ modules/portal/portal-soap-extender-test/bnd-api.bnd
@@ -1,7 +1,8 @@
-Bundle-Activator: com.liferay.portal.soap.extender.test.activators.simple.JaxwsApiBundleActivator
+Bundle-Activator: com.liferay.portal.soap.extender.test.activator.JaxwsApiBundleActivator
 Bundle-Name: Liferay Portal SOAP Extender Test API
 Bundle-SymbolicName: com.liferay.portal.soap.extender.test.api
 Bundle-Version: 1.0.0
 Include-Resource:\
-	com/liferay/portal/soap/extender/test/activators/simple=classes/com/liferay/portal/soap/extender/test/activators/simple,\
+	com/liferay/portal/soap/extender/test/activator=classes/com/liferay/portal/soap/extender/test/activator,\
+	com/liferay/portal/soap/extender/test/handler=classes/com/liferay/portal/soap/extender/test/handler,\
 	com/liferay/portal/soap/extender/test/service=classes/com/liferay/portal/soap/extender/test/service
