+++ modules/portal/portal-soap-extender-test/bnd-component-handler.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Portal SOAP Extender Test Component Handler
+Bundle-SymbolicName: com.liferay.portal.soap.extender.test.component.handler
+Bundle-Version: 1.0.0
+Include-Resource:\
+	com/liferay/portal/soap/extender/test/handler=classes/com/liferay/portal/soap/extender/test/handler,\
+	com/liferay/portal/soap/extender/test/service=classes/com/liferay/portal/soap/extender/test/service
