+++ modules/portal/portal-soap-extender-test/bnd-component.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Portal SOAP Extender Test Simple
-Bundle-SymbolicName: com.liferay.portal.soap.extender.test.simple
+Bundle-Name: Liferay Portal SOAP Extender Test Component
+Bundle-SymbolicName: com.liferay.portal.soap.extender.test.component
 Bundle-Version: 1.0.0
 Include-Resource: com/liferay/portal/soap/extender/test/service=classes/com/liferay/portal/soap/extender/test/service
