--- modules/portal/portal-soap-extender-test/bnd-handlers-api.bnd
@@ -1,8 +1,0 @@
-Bundle-Activator: com.liferay.portal.soap.extender.test.activators.handlers.JaxwsApiBundleActivatorWithHandlers
-Bundle-Name: Liferay Portal SOAP Extender Test API With Handlers
-Bundle-SymbolicName: com.liferay.portal.soap.extender.test.api.handlers
-Bundle-Version: 1.0.0
-Include-Resource:\
-	com/liferay/portal/soap/extender/test/activators/handlers=classes/com/liferay/portal/soap/extender/test/activators/handlers,\
-	com/liferay/portal/soap/extender/test/handlers=classes/com/liferay/portal/soap/extender/test/handlers,\
-	com/liferay/portal/soap/extender/test/service=classes/com/liferay/portal/soap/extender/test/service
