--- modules/portal/portal-soap-extender-test/bnd-handlers.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Portal SOAP Extender Test Simple With Handlers
-Bundle-SymbolicName: com.liferay.portal.soap.extender.test.simple.handlers
-Bundle-Version: 1.0.0
-Include-Resource:\
-	com/liferay/portal/soap/extender/test/handlers=classes/com/liferay/portal/soap/extender/test/handlers,\
-	com/liferay/portal/soap/extender/test/service=classes/com/liferay/portal/soap/extender/test/service
